requirements.txt
discord.py>=2.3.0
pyfiglet>=0.8.0
parsedatetime>=2.6
aiofiles>=23.1.0
python-dotenv>=1.0.0